Output에서 DrinkFactory로 변경, 음료 생성 시 사용할 enum 추가
