+++ VendingMachineApp/VendingMachineApp/Model/Productization/enum.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum KinkOfDrink: String {
+    case blueBottleColdBrew = "BlueBottle Cold Brew"
+    case starBucksColdBrew = "StarBucks Cold Brew"
+    case seoulStrawberryMilk = "서울 우유 딸기맛"
+    case bingBananaMilk = "빙그레 바나나우유"
+    case sanpellegrino = "Sanpellegrino"
+    case fanta = "Fanta"
+}
